requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachelib==0.13.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
loader==2017.9.11
Logbook==1.7.0.post0
MarkupSafe==2.1.5
msgspec==0.18.6
multidict==6.0.5
pendulum==3.0.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.5
requests==2.32.3
rq==1.16.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
termcc==2019.8.2
terminaltables==3.1.10
time-machine==2.14.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
worker==2018.4.1
yarl==1.9.4
